Modified to make Module::Install work on checkout
Module::Install is shipped with the dist, but it also needs to be
shipped in the Git repo otherwise you can't actually run Makefile.PL
and tests.
So I removed it from .gitignore (it's also already in MANIFEST), and
checked it in. This allows me to just clone the module from git on
a new box and have it work.